requirements.txt
# Web Framework
fastapi>=0.115.2,<1.0.0
uvicorn>=0.30.0,<1.0.0

# Database
sqlalchemy>=2.0.38,<3.0.0
alembic>=1.13.0,<2.0.0

# Environment Variables
python-dotenv>=1.0.1,<2.0.0

# LangFlow and LiteLLM
langflow==1.2.0
litellm==1.60.2
langchain>=0.1.0,<1.0.0
langchain-openai>=0.0.5,<1.0.0

# Testing
pytest==7.4.2
pytest-cov==2.12.1
httpx>=0.27.0

# Utilities
requests>=2.31.0
pydantic>=2.6.4
loguru>=0.7.1,<1.0.0 